ExtractTextAPI/requirements.txt
asgiref==3.7.1
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
comtypes==1.1.14
cx-Freeze==6.15.0
cx-Logging==3.1.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.9
django-cors-headers==4.1.0
django-mssql==1.8
django-mssql-backend==2.8.1
djangorestframework==3.14.0
dnspython==2.3.0
executing==1.2.0
filelock==3.12.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
image==1.5.33
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
lief==0.13.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mssql==1.0.1
mssql-django==1.3
nest-asyncio==1.5.6
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pyzmq==25.0.2
regex==2023.5.5
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.17
sqlparse==0.2.4
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
wcwidth==0.2.6
Werkzeug==2.3.4
